Div should be double!

+++ SWT-TestedProject/Calculator.cs
@@ -24,7 +24,7 @@ namespace SWT_TestedProject
             return a * b;
         }
 
-        public int Divide(int a, int b)
+        public double Divide(double a, double b)
         {
             return a / b;
         }
